+++ server/src/Lib/Server/Events/Workspace/index.ts
@@ -1,6 +1,6 @@
 /*	Auto generated	*/
 
-export * from "./OnDidCreateFilesAsync";
-export * from "./OnDidDeleteFilesAsync";
-export * from "./OnDidRenameFilesAsync";
-export * from "./OnWorkspaceFolderChange";
+export * from "./Create";
+export * from "./Delete";
+export * from "./Rename";
+export * from "./Change";
